--- VideoStreaming/Base.lproj/Main.storyboard
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14868" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="32c-Z1-l5F">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14824"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--View Controller-->
-        <scene sceneID="49a-7k-uFk">
-            <objects>
-                <viewController id="32c-Z1-l5F" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="lfy-Mm-9tv">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <containerView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="4Hd-1X-BIz">
-                                <rect key="frame" x="0.0" y="44" width="414" height="818"/>
-                                <constraints>
-                                    <constraint firstAttribute="width" constant="414" id="UdD-7y-K3A"/>
-                                    <constraint firstAttribute="height" constant="818" id="bq9-Zq-NPt"/>
-                                </constraints>
-                                <connections>
-                                    <segue destination="G8d-iX-Bah" kind="embed" identifier="videoPlaylistViewController" id="hTq-9R-vMV"/>
-                                </connections>
-                            </containerView>
-                        </subviews>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                        <constraints>
-                            <constraint firstItem="4Hd-1X-BIz" firstAttribute="trailing" secondItem="Z4X-I9-7CE" secondAttribute="trailing" id="72U-Uc-cVq"/>
-                            <constraint firstItem="4Hd-1X-BIz" firstAttribute="leading" secondItem="Z4X-I9-7CE" secondAttribute="leading" id="hrK-xd-GEu"/>
-                            <constraint firstItem="4Hd-1X-BIz" firstAttribute="bottom" secondItem="Z4X-I9-7CE" secondAttribute="bottom" id="ouZ-HP-Tjs"/>
-                            <constraint firstItem="4Hd-1X-BIz" firstAttribute="top" secondItem="Z4X-I9-7CE" secondAttribute="top" id="qhV-px-Jbi"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="Z4X-I9-7CE"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="56V-VE-kDB" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-320.28985507246381" y="-141.29464285714286"/>
-        </scene>
-        <!--Video Playlist View Controller-->
-        <scene sceneID="ctM-H4-t62">
-            <objects>
-                <viewController id="G8d-iX-Bah" customClass="VideoPlaylistViewController" customModule="VideoStreaming" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="dIz-tg-IT4">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="818"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="Dfm-uC-jnc" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="633" y="-430"/>
-        </scene>
-    </scenes>
-</document>
