--- VideoStreaming/MainViewController.swift
@@ -1,17 +1,0 @@
-//
-//  MainViewController.swift
-//  VideoStreaming
-//
-//  Created by Julian Llorensi on 24/09/2019.
-//  Copyright © 2019 ESPN. All rights reserved.
-//
-
-import UIKit
-
-class MainViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-    
-}
