+++ VideoStreaming/Resources/Assets.xcassets/Contents.json
