+++ VideoStreaming/Resources/Assets.xcassets/share_image.imageset/Contents.json
