+++ VideoStreamingTests/VideoPlaylistDataSourceTests.swift
@@ -9,19 +9,6 @@
 import XCTest
 @testable import VideoStreaming
 
-class MockGameStreamsAPI: GameStreamsAPIProtocol {
-    var fetchCompletionResult: Swift.Result<[VideoStreaming.Stream],APIError>
-    
-    init(result: Swift.Result<[VideoStreaming.Stream],APIError>) {
-        self.fetchCompletionResult = result
-    }
-    
-    func fetchGameStreams(ofGame gameId: String, completion: @escaping (Result<[VideoStreaming.Stream], APIError>) -> Void) {
-        DispatchQueue.main.async {
-            completion(self.fetchCompletionResult)
-        }
-    }
-}
 
 class VideoPlaylistDataSourceTests: XCTestCase {
