+++ VideoStreamingTests/VideoStreamingTests.swift
@@ -30,5 +30,4 @@ class VideoStreamingTests: XCTestCase {
             // Put the code you want to measure the time of here.
         }
     }
-
 }
